make neural network serialization less vocabulary-dependent
the old serialization function just stored the whole array,
without any additional information of what vector belonged to which word.
changes in the vocabulary rendered the whole data useless.
the new format stores vectors for each word in vocabulary.
this way the trained neural network can be used with vocabularies that
grew or shrunk.